instead of adding tests to integ-test-zip add a smoke test qa project for http tests
