+++ qa/smoke-test-http/src/test/java/org/elasticsearch/http/HttpCompressionIT.java
