+++ settings.gradle
@@ -56,6 +56,7 @@ List projects = [
   'qa:smoke-test-multinode',
   'qa:smoke-test-plugins',
   'qa:smoke-test-reindex-with-painless',
+  'qa:smoke-test-http',
   'qa:vagrant',
 ]
